move TestLoad to persist_test
